+++ index.html
@@ -9,28 +9,26 @@
   </head>
 
   <body>
-    <div id="app">
-      <!-- GNB -->
-      <lunch-header></lunch-header>
-      <!-- 카테고리/정렬 필터 -->
-      <section class="restaurant-filter-container">
-        <filter-list
-          name="category"
-          id="categoryFilter"
-          options="전체,한식,중식,일식,양식,아시안,기타"
-        ></filter-list>
-        <filter-list
-          name="sorting"
-          id="sortingFilter"
-          options="이름순,거리순"
-        ></filter-list>
-      </section>
-      <section class="restaurant-list-container">
-        <ul class="restaurant-list">
-          <restaurant-box></restaurant-box>
-        </ul>
-      </section>
-      <restaurant-modal></restaurant-modal>
-    </div>
+    <!-- GNB -->
+    <lunch-header></lunch-header>
+    <!-- 카테고리/정렬 필터 -->
+    <section class="restaurant-filter-container">
+      <filter-list
+        name="category"
+        id="categoryFilter"
+        options="전체,한식,중식,일식,양식,아시안,기타"
+      ></filter-list>
+      <filter-list
+        name="sorting"
+        id="sortingFilter"
+        options="이름순,거리순"
+      ></filter-list>
+    </section>
+    <section class="restaurant-list-container">
+      <ul class="restaurant-list">
+        <restaurant-box></restaurant-box>
+      </ul>
+    </section>
+    <restaurant-modal></restaurant-modal>
   </body>
 </html>
