+++ src/components/Modal/index.css
@@ -17,8 +17,9 @@
 
 .modal-container {
   position: fixed;
-  width: 100%;
+  right: 0;
   bottom: 0;
+  left: 0;
   padding: 32px 16px;
   border-radius: 8px 8px 0px 0px;
   background: var(--grey-100);
@@ -30,4 +31,5 @@
 
 .button-container {
   display: flex;
+  justify-content: space-between;
 }
