+++ src/components/SelectList/index.css
@@ -22,28 +22,18 @@
 }
 
 .form-item input,
-.form-item textarea,
 .form-item select {
   padding: 8px;
   margin: 6px 0;
-
   border: 1px solid var(--grey-200);
   border-radius: 8px;
-
   font-size: 16px;
-}
-
-.form-item textarea {
-  resize: none;
 }
 
 .form-item select {
   height: 44px;
-
   padding: 8px;
-
   border: 1px solid var(--grey-200);
   border-radius: 8px;
-
   color: var(--grey-300);
 }
