+++ src/styles/index.css
@@ -3,16 +3,6 @@
 body {
   font-family: sans-serif;
   font-size: 16px;
-  max-width: 390px;
-  position: absolute;
-  left: 0;
-  right: 0;
-  margin: 0 auto;
-}
-
-.app {
-  display: flex;
-  flex-direction: column;
 }
 
 :root {
@@ -31,6 +21,31 @@ body {
 
   padding: 0 16px;
   margin-top: 24px;
+}
+
+/* Typography *************************************/
+.text-title {
+  font-size: 20px;
+  line-height: 24px;
+  font-weight: 600;
+}
+
+.text-subtitle {
+  font-size: 18px;
+  line-height: 28px;
+  font-weight: 600;
+}
+
+.text-body {
+  font-size: 16px;
+  line-height: 24px;
+  font-weight: 400;
+}
+
+.text-caption {
+  font-size: 14px;
+  line-height: 20px;
+  font-weight: 400;
 }
 
 /* 음식점 목록 */
